+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added support `text/uri-list` to `ObsidianPasteImg`.
 
+### Changed
+
+- `ObsidianToggleCheckbox` now works in visual mode for multiline toggle
+
 ## [v3.10.0](https://github.com/obsidian-nvim/obsidian.nvim/releases/tag/v3.10.0) - 2025-04-12
 
 ### Added